Упростил main
